Create `/profile` router
