+++ server/server.js
@@ -14,6 +14,7 @@ const cors = require("cors");
 const { mongoose } = require("./db/mongoose");
 const auth = require("./routes/auth");
 const blog = require("./routes/blog");
+const profile = require("./routes/profile");
 
 const app = express();
 app.use(cors());
@@ -22,6 +23,7 @@ app.use(bodyParser.json());
 app.use(express.static(publicPath));
 app.use("/auth", auth);
 app.use("/blog", blog);
+app.use("/profile", profile);
 
 app.listen(process.env.PORT, () => {
   console.log(`Server is live on port ${process.env.PORT}`);
